+++ tests/testthat/test-outputs.R
@@ -589,7 +589,7 @@ test_that("can generate summary report from zip file", {
 test_that("can generate comparison report with only 1 survey chosen", {
   ## Create a model output with only 1 option chosen for survey_prevalence
   model_output <- a_hintr_output_calibrated$model_output_path
-  output <- qs::qread(model_output)
+  output <- read_hintr_output(model_output)
   options <- yaml::read_yaml(text = output$info$options.yml)
   options$survey_prevalence <- options$survey_prevalence[1]
   output$info$options.yml <- yaml::as.yaml(options)
@@ -681,7 +681,7 @@ test_that("prevalence survey plots not drawn when using aggregate survey", {
 test_that("can generate comparison report with ANC data at T1 not macthed to model T1", {
   ## Create a model output with only 1 option chosen for survey_prevalence
   model_output <- a_hintr_output_calibrated$model_output_path
-  output <- qs::qread(model_output)
+  output <- read_hintr_output(model_output)
   options <- yaml::read_yaml(text = output$info$options.yml)
   options$anc_prevalence_year1 <- "2017"
   options$anc_art_coverage_year1 <- "2017"